+++ src/main/webapp/WEB-INF/templates/pebble/base.html
@@ -122,7 +122,8 @@
     <!-- IE10 viewport hack for Surface/desktop Windows 8 bug -->
     <!-- <script src="../../assets/js/ie10-viewport-bug-workaround.js"></script>  -->
     
-    <!-- Local script file for the template manager app -->
+    <!-- Local script files for the template manager app -->
+    <script src="../assets/js/tempss-constraints.js"></script>
     <script src="../assets/js/tempss-manager.js"></script>
     <script type="text/javascript">
 	$(document).ready( function() {
