+++ src/main/webapp/WEB-INF/templates/pebble/constraint_info.html
@@ -41,7 +41,7 @@
         <thead>
           <th>Constraint Var 1</th>
           <th>Constraint Var 2</th>
-          <th>Value Mappings</th>
+          <th>Value Mappings (Var 1 -> Var 2)</th>
         </thead>
         <tbody>
           {% for item in constraints %}
