+++ src/main/webapp/WEB-INF/templates/pebble/index.html
@@ -513,7 +513,7 @@
         
         // Handle click of constraint info link
         $('body').on('click', '.constraint-info-link', function(e) {
-        	showConstraintInfo(e);
+        	constraints.showConstraintInfo(e);
         });
 	});
 </script>
