+++ src/main/webapp/assets/js/tempss-constraints.js
@@ -1,0 +1,48 @@
+window.constraints = {
+	
+	setup: function(data, $nameNode, $treeRoot) {
+		log("Request to setup constraints for template...");
+		if(!data.constraints) {
+			log("There is no constraints information in the provided data object.");
+			return;
+		}
+		// Add a comment to the root node with a link to display the constraint
+		// information
+    	var $constraintHtml = $('<div class="constraint-header">' + 
+    		'<i class="glyphicon glyphicon-link"></i> This template ' +
+    		'has constraints set. Click <a href="#" ' + 
+    		'class="constraint-info-link">here</a> for details.</div>');
+    	$constraintHtml.insertAfter($nameNode);
+    	
+    	// Add a constraint icon to each template node involved in a 
+    	// constraint relationship.
+    	
+	},
+
+	// Display a modal showing constraint information for the current template.
+	showConstraintInfo: function(e) {
+		var $target = $(e.currentTarget);
+		var templateName = $target.parent().parent().children('span[data-fqname]').data('fqname');
+		var templateId = $('input[name="componentname"]').val();
+		log("Constraint info requested for solver template <" + templateName + "> with ID <" + templateId + ">...");
+		e.preventDefault();
+	
+		// Get the constraint info and display in a BootstrapDialog
+		BootstrapDialog.show({
+			title: "Constraint information",
+	        message: function(dialog) { 
+	        	var $message = $('<div></div>');
+	        	
+	        	// Try to load the constraint data and show an error if loading fails
+	        	$message.load('/tempss/api/constraints/' + templateId, 
+	        			function( response, status, xhr ) {
+	        				if (status=="error") {
+	        					$message.html('<div class="alert alert-danger"><b>Unable to access constraint information.</b> An error has occurred accessing the constraint data from TemPSS for this template.</div>');
+	        				}
+	      		});
+	        	return $message
+	        }
+	    });
+	}
+
+}
